chore: DRY `ggml.rs` + `gguf.rs` common adapter config logic
Finally, all this extra boilerplate can be shifted into the model config `Adapter` struct to self-contain in a single method.

This required adjusting ownership a little to satisfy the compiler. The original `from_gguf()` and `from_ggml()` methods are unaffected, they still receive the expected params as reference.
