+++ mistralrs-core/src/pipeline/gguf.rs
@@ -378,40 +378,11 @@ impl Loader for GGUFLoader {
             ModelKind::XLoraGGUF => {
                 let xlora_paths = vec![paths.get_classifier_path().as_ref().unwrap().to_path_buf()];
                 let xlora_config = Some(paths.get_classifier_config().as_ref().unwrap().clone());
-                // Common:
-                let lora_config = paths.get_adapter_configs().as_ref().unwrap();
-                let ordering = paths.get_ordering().as_ref().unwrap();
-                let preload_adapters = &load_preload_adapters(
-                    paths.get_lora_preload_adapter_info(),
-                    DType::F32,
-                    device,
-                    silent,
-                )?;
-
-                let vb = &from_mmaped_safetensors(
-                    xlora_paths,
-                    paths
-                        .get_adapter_filenames()
-                        .as_ref()
-                        .unwrap()
-                        .iter()
-                        .map(|(_, x)| (*x).to_owned())
-                        .collect::<Vec<_>>(),
-                    DType::F32,
-                    device,
-                    silent,
-                )?;
 
                 let model_config = (
                     ModelConfig::FileGGUF { ct: model, reader: &mut file },
                     ModelConfig::Device { device, mapper },
-                    ModelConfig::Adapter {
-                        lora_config,
-                        xlora_config,
-                        vb,
-                        ordering,
-                        preload_adapters,
-                    },
+                    ModelConfig::Adapter::try_new(paths, device, silent, xlora_paths, xlora_config)?,
                 );
 
                 match arch {
@@ -424,40 +395,11 @@ impl Loader for GGUFLoader {
                 is_lora = true;
                 let xlora_paths = vec![];
                 let xlora_config = None;
-                // Common:
-                let lora_config = paths.get_adapter_configs().as_ref().unwrap();
-                let ordering = paths.get_ordering().as_ref().unwrap();
-                let preload_adapters = &load_preload_adapters(
-                    paths.get_lora_preload_adapter_info(),
-                    DType::F32,
-                    device,
-                    silent,
-                )?;
-
-                let vb = &from_mmaped_safetensors(
-                    xlora_paths,
-                    paths
-                        .get_adapter_filenames()
-                        .as_ref()
-                        .unwrap()
-                        .iter()
-                        .map(|(_, x)| (*x).to_owned())
-                        .collect::<Vec<_>>(),
-                    DType::F32,
-                    device,
-                    silent,
-                )?;
 
                 let model_config = (
                     ModelConfig::FileGGUF { ct: model, reader: &mut file },
                     ModelConfig::Device { device, mapper },
-                    ModelConfig::Adapter {
-                        lora_config,
-                        xlora_config,
-                        vb,
-                        ordering,
-                        preload_adapters,
-                    },
+                    ModelConfig::Adapter::try_new(paths, device, silent, xlora_paths, xlora_config)?,
                 );
 
                 match arch {
