+++ mistralrs-core/src/xlora_models/quantized_phi3.rs
@@ -235,11 +235,11 @@ impl ModelWeights {
             reader,
             device,
             lora_config,
-            vb,
+            &vb,
             ordering,
             xlora_config,
             mapper,
-            preload_adapters,
+            &preload_adapters,
         )
     }
 }
